removed mock data and fixed file structure
