--- frontend/admin-crm/src/pages/settings/communication/index.ts
@@ -1,5 +1,0 @@
-// frontend/admin-crm/src/pages/settings/communication/index.ts
-
-export { CommunicationTemplates } from './CommunicationTemplates';
-export { CommunicationRecords } from './CommunicationRecords';
-export { CommunicationNotifications } from './Notifications';
